Added exclude patterns support for Clean Task
* Added test file

* Refactoring and cosmetic fixes
